pyproject.toml
[project]
name = "florae-demo-with-appium"
version = "0.1.0"
description = ""
authors = [
    { name = "Dominic", email = "127579510+Tanakrit-D@users.noreply.github.com" }
]
dependencies = [
    "appium-python-client>=4.2.0",
    "allure-pytest>=2.13.5",
    "pylint-pytest>=1.1.8",
    "appium-swipe-actions>=0.1.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "pytest<=8.2.0",
    "ruff>=0.6.2",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D100","D212","S101","TCH001","TCH002","TCH003"]

[tool.ruff]
line-length = 140
